[zephyr] update to tag v4.2.1
* update to tag v4.2.1

    commit 64e98eb5278d790f3b3e5279ca8b39b45097bbc0 (tag: v4.2.1)
    Author: Daniel DeGrasse <ddegrasse@tenstorrent.com>
    Date:   Fri Oct 3 14:02:58 2025 -0500

        release: Zephyr v4.2.1

        Set the version to v4.2.1


DEPEND: I3ddc4f29a37732ab469ace6239a6942390ee482c

+++ drivers/wifi/ameba/src/ameba_wifi_drv.c
@@ -120,7 +120,7 @@ void dhcpv4_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event, str
 }
 
 /* called in mgmt_thread in net_mgmt */
-static void wifi_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void wifi_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 			       struct net_if *iface)
 {
 	/* zephyr wifi todo, temp */
